+++ superuser/example/lib/main.dart
@@ -58,6 +58,38 @@ class Context extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    void onDisplayingGroups() async {
+      await showDialog<void>(
+          context: context,
+          barrierDismissible: false,
+          builder: (context) {
+            List<String> gpList = Superuser.groups.toList(growable: false);
+
+            return AlertDialog(
+              title: const Text("Groups"),
+              content:
+                  Column(mainAxisSize: MainAxisSize.min, children: <Widget>[
+                Text("Total joined groups: ${gpList.length}"),
+                SizedBox(
+                    width: 400,
+                    height: 275,
+                    child: ListView.builder(
+                        shrinkWrap: true,
+                        itemCount: gpList.length,
+                        itemBuilder: (context, index) =>
+                            ListTile(title: Text(gpList[index]))))
+              ]),
+              actions: <TextButton>[
+                TextButton(
+                    onPressed: () {
+                      Navigator.pop(context);
+                    },
+                    child: const Text("OK"))
+              ],
+            );
+          });
+    }
+
     return Scaffold(
         appBar: AppBar(title: const Text("Superuser")),
         drawer: Drawer(
@@ -121,7 +153,13 @@ class Context extends StatelessWidget {
                   child: ListTile(
                       title: const Text("Run as superuser", style: _titleStyle),
                       trailing: Text(Superuser.isActivated ? "Yes" : "No",
-                          style: _valueStyle)))
+                          style: _valueStyle))),
+              const Divider(),
+              ListTile(
+                  title: const Text("Group"),
+                  trailing: ElevatedButton(
+                      onPressed: onDisplayingGroups,
+                      child: const Text("List all joined groups")))
             ]));
   }
 }