+++ superuser_plugin_unix/lib/superuser_plugin_unix.dart
@@ -28,11 +28,35 @@ final class UnixSuperuser extends SuperuserPlatform {
         });
 
   @override
-  bool get isActivated => isSuperuser;
+  bool get isActivated =>
+      onGettingProperties((lib) => SuperuserPluginUnixBindings(lib).is_root());
 
   @override
-  bool get isSuperuser =>
-      onGettingProperties((lib) => SuperuserPluginUnixBindings(lib).is_root());
+  bool get isSuperuser {
+    if (isActivated) {
+      return true;
+    }
+
+    return onGettingProperties((lib) {
+      final SuperuserPluginUnixBindings bindings =
+          SuperuserPluginUnixBindings(lib);
+
+      Pointer<Bool> result = ffi.calloc<Bool>();
+
+      try {
+        int errCode = bindings.is_sudo_group(result);
+
+        if (errCode > 0) {
+          throw SuperuserProcessError(
+              errCode, "Unable to retrive group information.");
+        }
+
+        return result.value;
+      } finally {
+        ffi.calloc.free(result);
+      }
+    });
+  }
 
   @override
   String get whoAmI => onGettingProperties((lib) {
@@ -44,7 +68,7 @@ final class UnixSuperuser extends SuperuserPlatform {
         try {
           int errCode = bindings.get_uname(resultPtr);
 
-          if (errCode != 0) {
+          if (errCode > 0) {
             throw SuperuserProcessError(errCode, "Unable to retrive username.");
           }
 
@@ -53,6 +77,43 @@ final class UnixSuperuser extends SuperuserPlatform {
           return result;
         } finally {
           ffi.calloc.free(resultPtr);
+        }
+      });
+
+  @override
+  Iterable<String> get groups => onGettingProperties((lib) sync* {
+        final SuperuserPluginUnixBindings bindings =
+            SuperuserPluginUnixBindings(lib);
+
+        Pointer<Pointer<gid_t>> gps = ffi.calloc<Pointer<gid_t>>();
+        Pointer<Int> size = ffi.calloc<Int>();
+
+        try {
+          int errCode = bindings.get_current_user_group(size, gps);
+          if (errCode > 0) {
+            throw SuperuserProcessError(
+                errCode, "Unable to obtain current user's associated groups.");
+          }
+
+          Pointer<gid_t> gids = gps.value;
+          Pointer<Pointer<Char>> gpName = ffi.calloc<Pointer<Char>>();
+
+          try {
+            for (int i = 0; i < size.value; i++) {
+              int nameErrCode = bindings.get_group_name_by_gid(gids[i], gpName);
+              if (nameErrCode > 0) {
+                throw SuperuserProcessError(
+                    nameErrCode, "Failed to list group name.");
+              }
+
+              yield gpName.value.cast<ffi.Utf8>().toDartString();
+            }
+          } finally {
+            ffi.calloc.free(gpName);
+            bindings.flush(gids.cast<Void>());
+          }
+        } finally {
+          [gps, size].forEach(ffi.calloc.free);
         }
       });
 }