+++ superuser_plugin_windows/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.0.0
+
+* Implement listing joined group for current user.
+
+## 2.0.0-m.2
+
+* Fix incorrect configuration in CMakeLists.txt
+
+## 2.0.0-m.1
+
+* Change condition of `isSuperuser` by finding local user joined Administrators group already.
+* Adjust buffer size to 256 wide character.
+
 ## 1.0.0
 
 * Add throw error if native code returns with error.